Add ApplyLog class for reading apply logs
This is a backwards incompatible change, as the "apply" and "destroy" return types have changed. Like the 'plan' function they now return a tuple that contains the raw results and a processed class.

The ApplyLog class reads through the streaming json terraform outputs. It can read the entire log at once to return final results, or it can have lines added to it as they come in for streaming results.